refactor: clean up feature/moodle branch before PR
Extract shared helpers (_to_dict, _HTML_TAG_RE, _CONFIRM_INSTRUCTIONS,
_EXPORTER_MAP) to eliminate duplicate patterns across client and skills.
Move LTI templates to lti_templates.py, consolidate test helpers into
conftest.py, remove 6 dead model classes and 3 orphan log constants.
Fix auth_disabled to only gate on OIDC so --no-auth-mode works with
LTI platforms configured.
