+++ src/soliplex/loggers.py
@@ -81,9 +81,6 @@ ROOM_UNKNOWN_CHUNK_ID = "unknown chunk id: %s"
 
 LTI_UNKNOWN_PLATFORM = "unknown LTI platform"
 LTI_INVALID_LAUNCH = "invalid LTI launch request"
-LTI_LOGIN = "LTI OIDC login initiated"
-LTI_LAUNCH = "LTI launch completed"
-LTI_VALIDATION_FAILED = "LTI id_token validation failed"
 LTI_SECRET_NOT_CONFIGURED = "LTI_SESSION_SECRET not configured"
