+++ src/soliplex/lti/__init__.py
@@ -1,0 +1,6 @@
+"""LTI 1.3 claim URI constants shared across the lti package."""
+
+LTI_CLAIM_ROLES = "https://purl.imsglobal.org/spec/lti/claim/roles"
+LTI_CLAIM_RESOURCE_LINK = (
+    "https://purl.imsglobal.org/spec/lti/claim/resource_link"
+)
