+++ tests/unit/test_installation.py
@@ -595,23 +595,14 @@ def test_installation_authorization_dburi_async():
     )
 
 
-@pytest.mark.parametrize(
-    "w_oidc_configs, w_lti_configs, expected",
-    [
-        ([], [], True),
-        ([object()], [], False),
-        ([], [object()], False),
-        ([object()], [object()], False),
-    ],
-)
-def test_installation_auth_disabled(w_oidc_configs, w_lti_configs, expected):
+@pytest.mark.parametrize("w_oidc_configs", [[], [object()]])
+def test_installation_auth_disabled(w_oidc_configs):
     i_config = mock.create_autospec(config_installation.InstallationConfig)
     i_config.oidc_auth_system_configs = w_oidc_configs
-    i_config.lti_platform_configs = w_lti_configs
 
     the_installation = installation.Installation(i_config)
 
-    assert the_installation.auth_disabled == expected
+    assert the_installation.auth_disabled == (not w_oidc_configs)
 
 
 def test_installation_oidc_auth_system_configs():
